+++ scripts/ingest-test-fixtures-update.sh
@@ -41,9 +41,10 @@ fi
 docker run --rm -v "$SCRIPT_DIR"/../unstructured:/root/unstructured -v \
    "$SCRIPT_DIR"/../test_unstructured_ingest:/root/test_unstructured_ingest \
    -w /root "$IMAGE_NAME" \
-   bash -c "export OVERWRITE_FIXTURES=true && source ~/.bashrc && pyenv activate unstructured &&
+   bash -c "export OVERWRITE_FIXTURES=true && source ~/.bashrc && pyenv activate unstructured && tesseract --version &&
                ./test_unstructured_ingest/test-ingest-azure.sh &&
                ./test_unstructured_ingest/test-ingest-github.sh &&
                ./test_unstructured_ingest/test-ingest-biomed-api.sh &&
                ./test_unstructured_ingest/test-ingest-biomed-path.sh &&
-               ./test_unstructured_ingest/test-ingest-s3.sh"
+               ./test_unstructured_ingest/test-ingest-s3.sh &&
+               ./test_unstructured_ingest/test-ingest-slack.sh"