+++ test_unstructured_ingest/test-ingest-against-api.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -e
+
 SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
 cd "$SCRIPT_DIR"/.. || exit 1
 
@@ -11,8 +13,10 @@ PYTHONPATH=. ./unstructured/ingest/main.py \
     --verbose \
     --reprocess
 
-if [ "$(find 'api-ingest-output' -type f -printf '.' | wc -c)" != 2 ]; then
+set +e
+
+if [ "$(find 'api-ingest-output' -type f -printf '.' | wc -c)" != 3 ]; then
    echo
-   echo "2 files should have been created."
+   echo "3 files should have been created."
    exit 1
 fi