+++ test_unstructured_ingest/test-ingest-biomed-path.sh
@@ -1,4 +1,12 @@
 #!/usr/bin/env bash
+# shellcheck disable=SC2317
+# NOTE(crag): remove above shellcheck line when the biomed issue is fixed
+echo "Skipping test-ingest-biomed-path.sh,"
+echo "see https://github.com/Unstructured-IO/unstructured/issues/468"
+echo
+exit 0
+
+set -e
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 cd "$SCRIPT_DIR"/.. || exit 1
@@ -21,12 +29,15 @@ PYTHONPATH=. ./unstructured/ingest/main.py \
     --biomed-path "oa_pdf/07/07/sbaa031.073.PMC7234218.pdf" \
     --structured-output-dir biomed-ingest-output-path \
     --num-processes 2 \
-    --verbose \
+    --reprocess \
     --download-dir biomed-download-path \
-    --preserve-downloads
+    --preserve-downloads \
+    --verbose
 
 
 OVERWRITE_FIXTURES=${OVERWRITE_FIXTURES:-false}
+
+set +e
 
 # to update ingest test fixtures, run scripts/ingest-test-fixtures-update.sh on x86_64
 if [[ "$OVERWRITE_FIXTURES" != "false" ]]; then