+++ unstructured/file_utils/filetype.py
@@ -183,6 +183,11 @@ def detect_filetype(
         extension = extension.lower()
         if os.path.isfile(_filename) and LIBMAGIC_AVAILABLE:
             mime_type = magic.from_file(filename or file_filename, mime=True)  # type: ignore
+            # NOTE(crag): for older versions of the OS libmagic package, such as is currently
+            # installed on the Unstructured docker image, .json files resolve to "text/plain"
+            # rather than "application/json". this corrects for that case.
+            if mime_type == "text/plain" and extension == ".json":
+                return FileType.JSON
         else:
             return EXT_TO_FILETYPE.get(extension.lower(), FileType.UNK)
     elif file is not None:
@@ -305,11 +310,11 @@ def _detect_filetype_from_octet_stream(file: IO) -> FileType:
         archive = zipfile.ZipFile(file)
 
         archive_filenames = [f.filename for f in archive.filelist]
-        if all([f in archive_filenames for f in EXPECTED_DOCX_FILES]):
+        if all(f in archive_filenames for f in EXPECTED_DOCX_FILES):
             return FileType.DOCX
-        elif all([f in archive_filenames for f in EXPECTED_XLSX_FILES]):
+        elif all(f in archive_filenames for f in EXPECTED_XLSX_FILES):
             return FileType.XLSX
-        elif all([f in archive_filenames for f in EXPECTED_PPTX_FILES]):
+        elif all(f in archive_filenames for f in EXPECTED_PPTX_FILES):
             return FileType.PPTX
 
     logger.warning("Could not detect the filetype from application/octet-stream MIME type.")